Move tests according to profile template
