+++ common-requirements.txt
@@ -15,4 +15,8 @@ sphinx-intl<1
 transifex-client
 
 # Profile
-ocdsextensionregistry>=0.0.8
+ocdsextensionregistry>=0.0.11
+
+# Tests
+pytest==2.8.7
+selenium==3.5.0
