+++ include/common.mk
@@ -1,5 +1,5 @@
 # Update this file from a profile with:
-# curl -O https://raw.githubusercontent.com/open-contracting/standard_profile_template/master/include/common.mk
+# curl https://raw.githubusercontent.com/open-contracting/standard_profile_template/master/include/common.mk -o include/common.mk
 
 # See https://github.com/datamade/data-making-guidelines
