+++ include/prologue.mk
@@ -1,5 +1,5 @@
 # Update this file from a profile with:
-# curl -O https://raw.githubusercontent.com/open-contracting/standard_profile_template/master/include/prologue.mk
+# curl https://raw.githubusercontent.com/open-contracting/standard_profile_template/master/include/prologue.mk -o include/prologue.mk
 
 # See http://clarkgrubb.com/makefile-style-guide#prologue
 MAKEFLAGS += --warn-undefined-variables
