+++ requirements.txt
@@ -6,6 +6,5 @@
 -e git+https://github.com/OpenDataServices/sphinxcontrib-jsonschema.git@75e4427b0f6f6a23dff9019aeb5d64833aab104a#egg=sphinxcontrib-jsonschema
 -e git+https://github.com/OpenDataServices/sphinxcontrib-opendataservices.git@fab0ff0167d32ec243d42f272e0e50766299c078#egg=sphinxcontrib-opendataservices
 
-pytest==2.8.7
-selenium==3.5.0
+# standard/schema/utils/fetch_currency_codelist.py
 lxml
