+++ standard/docs/en/schema/changelog.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.5]
+
+### Documentation
+
+* Build the documentation [#875](https://github.com/open-contracting/standard/pull/875) [#880](https://github.com/open-contracting/standard/pull/880) [#886](https://github.com/open-contracting/standard/pull/886) [#889](https://github.com/open-contracting/standard/pull/889).
+
 ## [1.1.4] - 2019-06-25
 
 ### Advisories
