+++ standard/tests/conftest.py
@@ -1,0 +1,49 @@
+# Update this file from a profile with:
+# curl https://raw.githubusercontent.com/open-contracting/standard_profile_template/master/tests/conftest.py -o tests/conftest.py # noqa
+import os
+from http.server import HTTPServer, SimpleHTTPRequestHandler
+from multiprocessing import Process
+
+import pytest
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+
+BROWSER = os.environ.get('BROWSER', 'ChromeHeadless')
+
+
+@pytest.fixture(scope="module")
+def browser(request):
+    if BROWSER == 'ChromeHeadless':
+        chrome_options = Options()
+        chrome_options.add_argument("--headless")
+        chrome_options.add_argument("--no-sandbox")
+        browser = webdriver.Chrome(chrome_options=chrome_options)
+    else:
+        browser = getattr(webdriver, BROWSER)()
+    browser.implicitly_wait(3)
+    request.addfinalizer(lambda: browser.quit())
+    return browser
+
+
+@pytest.fixture(scope="module")
+def server(request):
+    if 'CUSTOM_SERVER_URL' in os.environ:
+        return os.environ['CUSTOM_SERVER_URL']
+
+    host = 'localhost'
+    port_number = 8331
+
+    def run():
+        os.chdir('build')
+        server_address = (host, port_number)
+        httpd = HTTPServer(server_address, SimpleHTTPRequestHandler)
+        httpd.serve_forever()
+
+    p = Process(target=run)
+    p.start()
+
+    def fin():
+        p.terminate()
+    request.addfinalizer(fin)
+
+    return 'http://{}:{}/'.format(host, port_number)
